+++ nnlo/optimize/coordinator.py
@@ -7,9 +7,9 @@ import time
 import hashlib
 import numpy as np
 from mpi4py import MPI
-from genetic_algorithm import GA
-from tag_lookup import tag_lookup
-from mpi_learn.logger import set_logging_prefix
+from .genetic_algorithm import GA
+from ..util.utils import opt_tag_lookup
+from ..util.logger import set_logging_prefix
 import logging
 
 class Coordinator(object):
@@ -153,7 +153,7 @@ class Coordinator(object):
         ## end all processes
         for proc in range(1, self.comm.Get_size()):
             logging.debug("Signaling process {} to exit".format(proc))
-            self.comm.send(None, dest=proc, tag=tag_lookup('params'))
+            self.comm.send(None, dest=proc, tag=opt_tag_lookup('params'))
         self.comm.Barrier()
         logging.info("Finished all iterations!")
         logging.info("Best parameters found: {} with value {}".format(self.best_params, self.best_fom))
@@ -213,5 +213,5 @@ class Coordinator(object):
         end = block_num * block_size 
         logging.debug("Launching block {}. Sending params to nodes from {} to {}".format(block_num, start,end))
         for proc in range(start, end+1):
-            self.comm.send(params, dest=proc, tag=tag_lookup('params')) 
-        self.req_dict[block_num] = self.comm.irecv(source=start, tag=tag_lookup('result'))
+            self.comm.send(params, dest=proc, tag=opt_tag_lookup('params')) 
+        self.req_dict[block_num] = self.comm.irecv(source=start, tag=opt_tag_lookup('result'))
